MemoryStream now uses 'unsigned int' instead of 'size_t' and gcc is happy.  size_t is a dangerous overload because you never know what built-in type will conflict.
git-svn-id: https://unittest-cpp.svn.sourceforge.net/svnroot/unittest-cpp@133 feadf2a4-a60d-0410-a25d-8df00a7d1702

+++ UnitTest++/src/MemoryOutStream.cpp
@@ -85,12 +85,11 @@ MemoryOutStream& MemoryOutStream::operator << (void const* p)
     return *this;    
 }
 
-MemoryOutStream& MemoryOutStream::operator << (size_t const s)
+MemoryOutStream& MemoryOutStream::operator << (unsigned int const s)
 {
     FormatToStream(*this, "%u", s);
     return *this;    
 }
-
 
 int MemoryOutStream::GetCapacity() const
 {
